Move filesystem implementation into broker core
Relocate the resolver engine, filesystem backends, 9P client, and provider adapter into litebox_broker_core while keeping the guest descriptor facade in litebox.


Copilot-Session: 239ec5f7-870a-4259-bcae-4ca85fb913a0
