+++ litebox/Cargo.toml
@@ -8,7 +8,6 @@ bitflags = "2.13.1"
 hashbrown = "0.15.2"
 smallvec = "1.13.2"
 spin = { version = "0.9.8", default-features = false, features = ["spin_mutex"] }
-tar-no-std = { version = "0.3.3", default-features = false, features = ["alloc"] }
 thiserror = { version = "2.0.6", default-features = false }
 zerocopy = { version = "0.8", default-features = false, features = ["derive"] }
 rangemap = { version = "1.5.1", features = ["const_fn"] }
@@ -16,6 +15,7 @@ buddy_system_allocator = { version = "0.11.0", default-features = false, feature
 # Depend on (currently unreleased) slabmalloc `main`, which contains some fixes on top of `0.11.0`
 slabmalloc = { git = "https://github.com/gz/rust-slabmalloc.git", rev = "19480b2e82704210abafe575fb9699184c1be110" }
 litebox_util_log = { version = "0.1.0", path = "../litebox_util_log" }
+litebox_broker_core = { version = "0.1.0", path = "../litebox_broker_core" }
 litebox_broker_local = { version = "0.1.0", path = "../litebox_broker_local" }
 litebox_broker_protocol = { version = "0.1.0", path = "../litebox_broker_protocol" }
 litebox_broker_transport = { version = "0.1.0", path = "../litebox_broker_transport" }
