+++ litebox_broker_core/Cargo.toml
@@ -7,7 +7,10 @@ edition = "2024"
 bitflags = { version = "2.9.0", default-features = false }
 hashbrown = "0.15.2"
 litebox_broker_protocol = { path = "../litebox_broker_protocol", version = "0.1.0" }
+litebox_platform = { path = "../litebox_platform", version = "0.1.0" }
+litebox_util_log = { path = "../litebox_util_log", version = "0.1.0" }
 spin = { version = "0.9.8", default-features = false, features = ["once", "rwlock", "spin_mutex"] }
+tar-no-std = { version = "0.3.3", default-features = false, features = ["alloc"] }
 thiserror = { version = "2.0.6", default-features = false }
 
 [lints]
