+++ litebox_broker_core/src/fs/nine_p/client.rs
@@ -9,11 +9,11 @@ use alloc::sync::Arc;
 use alloc::vec::Vec;
 use core::sync::atomic::{AtomicBool, AtomicU16, Ordering};
 
-use crate::utils::id_pool::IdPool;
 use litebox_platform::sync::{Mutex, RawSyncPrimitivesProvider};
 
 use super::Error;
 use super::fcall::{self, Fcall, FcallStr, GetattrMask, TaggedFcall};
+use super::id_pool::IdPool;
 use super::transport::{self, Read, Write};
 
 /// Pool of 9P Fid values
