+++ litebox_shim_windows/src/syscalls/registry.rs
@@ -58,8 +58,7 @@ use crate::nt_types::{
     read_unicode_string_at,
 };
 
-type RegistryFileSystem<Platform> =
-    litebox::fs::resolver::Resolver<Platform, litebox::fs::composer::Composer>;
+type RegistryFileSystem<Platform> = litebox::fs::resolver::Resolver<Platform>;
 
 pub(crate) struct RegistryKeySubsystem<Platform>(PhantomData<fn(Platform)>);
